requirements.txt
anthropic>=0.40.0
numpy>=1.24.0
pandas
pytest
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
tabulate
